dma-debug: fix printk formats (i386)
Fix printk format warnings in dma-debug:

  lib/dma-debug.c:645: warning: format '%016llx' expects type 'long long unsigned int', but argument 6 has type 'dma_addr_t'
  lib/dma-debug.c:662: warning: format '%016llx' expects type 'long long unsigned int', but argument 6 has type 'dma_addr_t'
  lib/dma-debug.c:676: warning: format '%016llx' expects type 'long long unsigned int', but argument 6 has type 'dma_addr_t'
  lib/dma-debug.c:686: warning: format '%016llx' expects type 'long long unsigned int', but argument 6 has type 'dma_addr_t'

+++ lib/dma-debug.c
@@ -649,7 +649,7 @@ static void check_sync(struct device *dev, dma_addr_t addr,
 		err_printk(dev, NULL, "DMA-API: device driver tries "
 				"to sync DMA memory it has not allocated "
 				"[device address=0x%016llx] [size=%llu bytes]\n",
-				addr, size);
+				(unsigned long long)addr, size);
 		goto out;
 	}
 
@@ -667,7 +667,7 @@ static void check_sync(struct device *dev, dma_addr_t addr,
 				"DMA memory with different direction "
 				"[device address=0x%016llx] [size=%llu bytes] "
 				"[mapped with %s] [synced with %s]\n",
-				addr, entry->size,
+				(unsigned long long)addr, entry->size,
 				dir2name[entry->direction],
 				dir2name[direction]);
 	}
@@ -681,7 +681,7 @@ static void check_sync(struct device *dev, dma_addr_t addr,
 				"device read-only DMA memory for cpu "
 				"[device address=0x%016llx] [size=%llu bytes] "
 				"[mapped with %s] [synced with %s]\n",
-				addr, entry->size,
+				(unsigned long long)addr, entry->size,
 				dir2name[entry->direction],
 				dir2name[direction]);
 
@@ -691,7 +691,7 @@ static void check_sync(struct device *dev, dma_addr_t addr,
 				"device write-only DMA memory to device "
 				"[device address=0x%016llx] [size=%llu bytes] "
 				"[mapped with %s] [synced with %s]\n",
-				addr, entry->size,
+				(unsigned long long)addr, entry->size,
 				dir2name[entry->direction],
 				dir2name[direction]);
